% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_symmetry_test.R
\name{power_symmetry_test}
\alias{power_symmetry_test}
\title{Compute empirical power using Monte Carlo methods}
\usage{
power_symmetry_test(
  data,
  statis = c("Bk", "Jk", "R", "Rs", "Mp"),
  Bk = 5,
  Jk = 6,
  type = "k",
  alpha = 0.05,
  nsize,
  rep,
  plot = TRUE,
  ncores = NULL
)
}
\arguments{
\item{data}{data frame or tribble with the case name and the four lambda parameters of the GLD.}

\item{statis}{Test statistic to be used. By default \code{statis = c("Bk", "Jk")}.
Available options are \code{"Bk"}, \code{"Bkc"}, \code{"Jk"}, \code{"R"}, \code{"Rs"}, \code{"Mp"}.}

\item{Bk}{Integer or vector of integers with the cut-off parameter(s) for the \eqn{B_k} statistic.
By default \code{Bk = 5}. Multiple values can be supplied, e.g. \code{Bk = c(5, 6, 7)}.}

\item{Jk}{Integer or vector of integers with the cut-off parameter(s) for the \eqn{J_k} statistic.
By default \code{Jk = 6}. Multiple values can be supplied, e.g. \code{Jk = c(4, 6, 8)}.}

\item{type}{Type of test. When the test is with known median, select \code{type = "k"};
use \code{type = "u"} for unknown median (estimated from the data).}

\item{alpha}{Significance level of the test. By default \eqn{\alpha = 0.05}.}

\item{nsize}{Sample size for each Monte Carlo replicate.}

\item{rep}{Number of replicates for the Monte Carlo process.}

\item{plot}{Logical. If \code{TRUE} (default), draws the empirical power function.}

\item{ncores}{Number of cores to use for parallel computation. By default \code{ncores = NULL},
which automatically uses all available cores minus one.}
}
\value{
A data frame with empirical powers for each test statistic and distribution.
}
\description{
This function computes the empirical power using Monte Carlo methods for a large number of
tests for symmetry with known median. To see the available tests, see details.
}
\details{
Computes empirical power on a tribble or data frame that contains the parameters of the
Generalised Lambda Distribution (GLD).
Jk: Corzo, J. and Babativa, G. (2013),
R: McWiliams, P. (1990),
Rs: Baklizi, A. (2003),
Mp: Modarres and Gastwirth (1996).
}
\examples{
\dontrun{
distributions <- tibble::tribble(
  ~case,     ~lambda1, ~lambda2, ~lambda3, ~lambda4,
  "Case 1A",        0, 0.197454, 0.134915, 0.134915
)
P20 <- power_symmetry_test(data = distributions, statis = c("Bk", "Jk"),
                           Bk = 5, Jk = 6, alpha = 0.05, nsize = 20, rep = 1000, plot = FALSE)
}
}
\references{
Corzo, J., & Babativa, G. (2013). A modified runs test for symmetry. Journal of Statistical
Computation and Simulation, 83(5), 984-991.
}
\author{
Giovany Babativa <jgbabativam@unal.edu.co>
}
