% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_test.R
\name{symmetry_test}
\alias{symmetry_test}
\title{Perform symmetry test}
\usage{
symmetry_test(
  x,
  statis = c("Bk", "Jk", "R", "Rs", "Mp", "Bkc", "Ck", "Gk"),
  type = "k",
  Bk = 5,
  Ck = 10,
  Jk = 6,
  Gk = 6,
  Gk_B = 9999,
  Mp = c(10, 20, 25),
  median = 0
)
}
\arguments{
\item{x}{vector of numeric information.}

\item{statis}{Test statistic to be used. By default \code{statis = c("Bk", "Jk")}.
Available options are \code{"Bk"}, \code{"Ck"}, \code{"Bkc"}, \code{"Jk"}, \code{"R"}, \code{"Rs"}, \code{"Mp"}, \code{"Gk"}.}

\item{type}{Type of test. When the test is with known median, select \code{type = "k"};
use \code{type = "u"} for unknown median (estimated from the data).}

\item{Bk}{Integer or vector of integers with the cut-off parameter(s) for the \eqn{B_k} statistic.
By default \code{Bk = 5}. Multiple values can be supplied, e.g. \code{Bk = c(5, 6, 7)}.}

\item{Ck}{Integer (or vector) with the cut-off(s) for the \eqn{C_k} statistic.
\eqn{C_k} is the conditional version of \eqn{B_k}: it uses the runs distribution
conditioned on \eqn{(n_{1k}, n_{0k})} among the \eqn{k} most extreme observations,
making it robust to misspecified medians. Requires \eqn{k \\geq 10} for reliable
size control. By default \code{Ck = 10}.}

\item{Jk}{Integer or vector of integers with the cut-off parameter(s) for the \eqn{J_k} statistic.
By default \code{Jk = 6}. Multiple values can be supplied, e.g. \code{Jk = c(4, 6, 8)}.}

\item{Gk}{Integer or vector of integers with the cut-off parameter(s) for the \eqn{G_k} statistic.
\eqn{G_k} is the Shannon entropy (in bits) of the run-length distribution in the tail segment
comprising the \eqn{k} observations of largest absolute value. Small values indicate that the
tail run-length distribution is concentrated (few distinct lengths), which is evidence of
asymmetry. The null distribution is calibrated by permutation of signs.
By default \code{Gk = 6}. Multiple values can be supplied, e.g. \code{Gk = c(5, 6, 7)}.}

\item{Gk_B}{Number of Monte Carlo permutations used to calibrate the null distribution of
\eqn{G_k}. By default \code{Gk_B = 9999}.}

\item{Mp}{Numeric vector with percentile trimming values (in percent) for the \eqn{M_p} statistic.
By default \code{Mp = c(10, 20, 25)}.}

\item{median}{The centre parameter around which to test symmetry. By default \code{median = 0}.}

\item{Bkc}{Integer with the cut-off parameter for the \eqn{B_{kc}} statistic, which uses
the conditional runs distribution. By default \code{Bkc = 11}.}
}
\value{
A data frame with statistical values and p.values
}
\description{
This function contains a large number of tests for symmetry with known median. To see the available tests, see
details.
}
\details{
Perform symmetry tests on numeric vectors or objects, choose test with vector stat.
Jk: Corzo, J. and Babativa, G. (2013),
R: McWiliams, P. (1990),
Rs: Baklizi, A. (2003)
Mp: Modarres and Gastwirth (1996).
}
\examples{
x <- rnorm(20)
#--- All tests
(test <- symmetry_test(x))
#--- Choose any test
(Jk_test <- symmetry_test(x, statis = "Jk", Jk = 6))
#--- Entropy-based test
(Gk_test <- symmetry_test(x, statis = "Gk", Gk = 6))
#--- Choose several tests with multiple cut-offs
(MyTest1 <- symmetry_test(x, statis = c("Bk", "Jk"), Bk = c(5, 7), Jk = c(6, 8)))
(MyTest2 <- symmetry_test(x, statis = c("Mp", "Jk", "Gk"), Jk = 6, Mp = c(10, 20, 25), Gk = c(5, 6)))
}
\references{
Corzo, J., & Babativa, G. (2013). A modified runs test for symmetry. Journal of Statistical Computation and Simulation, 83(5), 984-991.
Corzo, J., & Babativa, G. (2025). An entropy-based runs test for the hypothesis of symmetry. Manuscript in preparation.
Shannon, C.E. (1948). A mathematical theory of communication. Bell System Technical Journal, 27, 379-423.
}
\author{
Giovany Babativa <jgbabativam@unal.edu.co>
}
