% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_symmetry_test.R
\name{power_symmetry_test}
\alias{power_symmetry_test}
\title{Compute empirical power using Monte Carlo methods}
\usage{
power_symmetry_test(
  data,
  statis = c("Bk", "Jk", "R", "Rs", "Mp"),
  Bk = 5,
  Jk = 6,
  type = "k",
  alpha = 0.05,
  nsize,
  rep,
  plot = TRUE
)
}
\arguments{
\item{data}{data frame or tribble with name and lambda parameters of the DLG.}

\item{statis}{Test statistic to be used. By default \code{stat = c("Bk", "Jk")}}

\item{type}{Type of test. When the test is with know median, select \code{type = "k"} else \code{type = "u"} for unknown median.}

\item{alpha}{Size of test. By default \eqn{\alpha = 0.05}}

\item{nsize}{Sample Size.}

\item{rep}{Number of replicates for the Monte Carlo process.}

\item{plot}{draw empirical power function with results.}
}
\value{
A data frame with empirical powers for the test
}
\description{
This function compute the empirical power using Monte Carlo methods for a large number of tests for symmetry with known median. To see the available tests, see
details.
}
\details{
Compute empirical power on tribble or data frame that contains the distributions of the DLG.
Jk: Corzo, J. and Babativa, G. (2013),
R: McWiliams, P. (1990),
Rs: Baklizi, A. (2003)
Mp: Modarres and Gastwirth (1996).
}
\examples{
x <- rnorm(20)
#--- All test
(test <- symmetry_test(x))
#--- Choose any test
(Jk_test <- symmetry_test(x, stat = "Jk", Jk = 6))
#--- Choose severals tests
(MyTest1 <- symmetry_test(x, stat = c("Bk", "Jk"), Bk = 5, Jk = 6))
(MyTest2 <- symmetry_test(x, stat = c("Mp", "Jk"), Jk = 6, Mp = c(10, 20, 25)))
}
\references{
Corzo, J., & Babativa, G. (2013). A modified runs test for symmetry. Journal of Statistical Computation and Simulation, 83(5), 984-991.
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
