% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{samples}
\alias{samples}
\title{Generate random samples from the Generalised Lambda Distribution (GLD)}
\usage{
samples(lambda1, lambda2, lambda3, lambda4, ni, B)
}
\arguments{
\item{lambda1}{Location parameter \eqn{\lambda_1} of the GLD.}

\item{lambda2}{Scale parameter \eqn{\lambda_2} of the GLD.}

\item{lambda3}{First shape parameter \eqn{\lambda_3} of the GLD.}

\item{lambda4}{Second shape parameter \eqn{\lambda_4} of the GLD.}

\item{ni}{Sample size per replicate.}

\item{B}{Number of Monte Carlo replicates.}
}
\value{
A numeric matrix of dimension \code{ni x B}, where \code{ni} is the sample size and
\code{B} is the number of replicates. Each column is one centred random sample.
}
\description{
Generates a matrix of random samples from the Generalised Lambda Distribution (GLD)
using the inverse CDF method. Each column corresponds to one Monte Carlo replicate,
centred at the theoretical median.
}
\details{
The samples are centred by subtracting the theoretical median of the GLD.
}
\examples{
\dontrun{
mat <- samples(lambda1 = 0, lambda2 = 0.1975, lambda3 = 0.1349, lambda4 = 0.1349,
               ni = 20, B = 100)
dim(mat)  # 20 x 100
}
}
