% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_test.R
\name{symmetry_test}
\alias{symmetry_test}
\title{Perform symmetry test}
\usage{
symmetry_test(
  x,
  statis = c("Bk", "Jk", "R", "Rs", "Mp", "Bkc"),
  type = "k",
  Bk = 5,
  Jk = 6,
  Bkc = 11,
  Mp = c(10, 20, 25),
  median = 0
)
}
\arguments{
\item{x}{vector of numeric information.}

\item{statis}{Test statistic to be used. By default \code{stat = c("Bk", "Jk")}}

\item{type}{Type of test. When the test is with know median, select \code{type = "k"} else \code{type = "u"} for unknown median.}

\item{Bk}{Parameter of \eqn{B_k} statistic. By default \code{Bk = 5}}

\item{Jk}{Parameter of \eqn{J_k} statistic. By default \code{Jk = 6}}

\item{median}{The centre parameter around which to test symmetry. By default \code{median=0}}
}
\value{
A data frame with statistical values and p.values
}
\description{
This function contains a large number of tests for symmetry with known median. To see the available tests, see
details.
}
\details{
Perform symmetry tests on numeric vectors or objects, choose test with vector stat.
Jk: Corzo, J. and Babativa, G. (2013),
R: McWiliams, P. (1990),
Rs: Baklizi, A. (2003)
Mp: Modarres and Gastwirth (1996).
}
\examples{
x <- rnorm(20)
#--- All test
(test <- symmetry_test(x))
#--- Choose any test
(Jk_test <- symmetry_test(x, statis = "Jk", Jk = 6))
#--- Choose severals tests
(MyTest1 <- symmetry_test(x, statis = c("Bk", "Jk"), Bk = 5, Jk = 6))
(MyTest2 <- symmetry_test(x, statis = c("Mp", "Jk"), Jk = 6, Mp = c(10, 20, 25)))
}
\references{
Corzo, J., & Babativa, G. (2013). A modified runs test for symmetry. Journal of Statistical Computation and Simulation, 83(5), 984-991.
}
\author{
Giovany Babativa <gbabativam@gmail.com>
}
